cuda: the fused Q prologue narrowed the rotated pair to bf16, narrowing twice
`af16_pack_qraw` applied the head's RMS scale, rotated the tail in f32, and then
rounded the result to bf16 before af16_pack narrowed it to f16.  Two narrowings
(8 mantissa bits, then 10) where there must be one: the standalone arm stores
that same f32 rotation straight into the f16 buffer, and dev has no such round.
1-2 bf16 ULP on every rotated element, on the PRODUCTION path only.

That one line was the whole reason production disagreed with the branch's own
standalone arm and with dev.  Measured on a clean build against ~/devref-clean:

    mine attn_out(0), no Qcur armed vs dev            100% BYTE-EXACT
    mine attn_out(0), no Qcur vs Qcur armed           100% BYTE-EXACT (neutral)
    fused-arm logits == standalone-arm logits         identical to 9 decimals
    layer sweep: layer 0 and layer 1 BYTE-EXACT, layer 2 the first divergence

Before this, layer 1's input already differed (1.3%) and the armed dump set
decided which arm ran -- arming `Qcur` switched production to the standalone
prologue, so every dump-armed comparison in s57-s79 was reading a different code
path than the one production runs.  2de728f3 fixed exactly this; 683e406d
reverted it with no rationale but "This reverts commit ..."; this restores it.

The aggregate pearson against dev DIPS (p1 0.9605 -> 0.9434, p3 0.9721 ->
0.9639, sky 0.9843 -> 0.9848): a wrong Q that brought the logits closer to dev
means a second, larger error of opposite sign was being cancelled by it.  The
fix is right -- byte-exact with dev at layer 0, and the dump dependence is gone
-- but parity improves only once the layer-2 defect goes too.  See rows/L218.md
s114; the next target is layer 2, the first compressed layer.

+++ src/cuda/pulsar_cuda_attn_f16.cu
@@ -274,8 +274,14 @@ __device__ static inline uint32_t af16_pack_qraw(
         rope_pair_rotate_core_dev(x0, x1, c0 - q_nope, qp.n_rot, rope_pos, 0,
                                   qp.freq_base, qp.freq_scale, qp.ext_factor,
                                   qp.attn_factor, corr0, corr1, &r0, &r1);
-        x0 = __bfloat162float(__float2bfloat16(r0));
-        x1 = __bfloat162float(__float2bfloat16(r1));
+        /* f32, NOT narrowed here: the standalone arm stores this same f32
+         * rotation straight into the f16 buffer, and af16_pack below is the
+         * single narrowing.  Rounding to bf16 first cost 1-2 bf16 ULP on every
+         * rotated element, which is what made production's Q disagree with the
+         * standalone arm and with dev (2de728f3; the bare revert 683e406d put it
+         * back and this restores it). */
+        x0 = r0;
+        x1 = r1;
     }
     return af16_pack(x0, x1);
 }
